This post is about analyzing the data of YouTube. This total analysis is performed using Apache Spark
This post will help you to understand how to handle data sets that does
not have proper structure and how to sort the output of reducer.

Data Set Description

Column 1: Video id of 11 characters.

Column 2: Uploader of the video.

Column 3: Interval between day of establishment of YouTube and the date
of uploading of the video.

Column 4: Category of the video.

Column 5: Length of the video.

Column 6: Number of views for the video.

Column 7: Rating on the video.

Column 8: Number of ratings given for the video

Column 9: Number of comments on the videos.

Column 10: Related video ids with the uploaded video.