In this example we will be  taking a famous Tv show dataset i.e., The Daily show and we will be performing analysis on the guests who came t
Dataset Description:

YEAR –  The year the episode aired

GoogleKnowlege_Occupation -Their occupation or office, according to
Google’s Knowledge Graph or, if they’re not in there, how Stewart
introduced them on the program.

Show – Air date of episode. Not unique, as some shows had more than one
guest

Group – A larger group designation for the occupation. For instance, us
senators, us presidents, and former presidents are all under
“politicians”

Raw_Guest_List – The person or list of people who appeared on the show,
according to Wikipedia. The GoogleKnowlege_Occupation only refers to one
of them in a given row.